refactor: initFile generation
There were a few places where the initFile endpoint was previously
returning 500. The primary goal was to ensure a response is always
returned, even when the miner is not fully operational.

To make this usage explicit, a DiagnosticsReport class has
been created that will never throw an exception. Exceptions
are stringified.

In the process, each type of test has been refactored into a
distinct Diagnostic. A DiagnosticReport is constructed with
an arbitrary set of Diagnostics. The result of each is appended
to the DiagnosticsReport dict.

The current keys used for the initFile are terse and confusing,
eg PF, OK, etc. In an attempt to migrate towards human friendly
names, each Diagnostic uses the legacy keys in addition
to new verbose_human_friendly_keys.